[ refactor ] Relocate + generalize some functions dealing with constraints
