+++ src/full/Agda/TypeChecking/MetaVars.hs
@@ -628,26 +628,6 @@ problemType (CheckProjAppToKnownPrincipalArg _ _ _ _ _ t _ _ _ _) = t -- The tar
 problemType (CheckLambda _ _ _ t     ) = t
 problemType (DoQuoteTerm _ _ t)        = t
 
--- | Eta expand metavariables listening on the current meta.
-etaExpandListeners :: MetaId -> TCM ()
-etaExpandListeners m = do
-  ls <- getMetaListeners m
-  clearMetaListeners m  -- we don't really have to do this
-  mapM_ wakeupListener ls
-
--- | Wake up a meta listener and let it do its thing
-wakeupListener :: Listener -> TCM ()
-  -- Andreas 2010-10-15: do not expand record mvars, lazyness needed for irrelevance
-wakeupListener (EtaExpand x)         = etaExpandMetaSafe x
-wakeupListener (CheckConstraint _ c) = do
-  reportSDoc "tc.meta.blocked" 20 $ "waking boxed constraint" <+> prettyTCM c
-  modifyAwakeConstraints (c:)
-  solveAwakeConstraints
-
--- | Do safe eta-expansions for meta (@SingletonRecords,Levels@).
-etaExpandMetaSafe :: (MonadMetaSolver m) => MetaId -> m ()
-etaExpandMetaSafe = etaExpandMeta [SingletonRecords,Levels]
-
 -- | Eta-expand a local meta-variable, if it is of the specified kind.
 --   Don't do anything if the meta-variable is a blocked term.
 etaExpandMetaTCM :: [MetaKind] -> MetaId -> TCM ()
